chore(examples): refactor workbox-cli example and add missing scripts and dependencies
