+++ examples/workbox-cli/globIgnores.ts
