+++ examples/workbox-cli/package.json
@@ -10,11 +10,15 @@
     "build": "pnpx tsx ./build.ts",
     "build-split": "pnpx tsx ./build-split.ts",
     "cli": "workbox-cli",
+    "cli:generate": "workbox-cli generate-sw.config.ts",
+    "cli:build-sw": "workbox-cli build-sw.config.ts",
+    "cli:inject": "workbox-cli inject-manifest.config.ts",
     "cli:manifest": "workbox-cli manifest.config.ts",
-    "cli:build-sw": "workbox-cli build-sw.config.ts"
+    "cli:self-destroy": "workbox-cli -c self-destroy-sw"
   },
   "dependencies": {
     "@composable-vite-pwa/workbox-build": "workspace:*",
+    "@composable-vite-pwa/workbox-cli": "workspace:*",
     "@composable-vite-pwa/workbox-swkit": "workspace:*"
   }
 }
