+++ examples/workbox-cli/workbox.config.ts
@@ -1,16 +1,13 @@
 import { defineCliOptions } from '@composable-vite-pwa/workbox-cli'
 import { runtimeCaching } from './cache.ts'
-import { globIgnores } from './glogIgnores.ts'
+import { globIgnores } from './globIgnores.ts'
 
-// `get-manifest` cannot be authored with the build-only `defineOptions`, so the
-// CLI ships `defineCliOptions`, widened to all four strategies.
-export default defineCliOptions('get-manifest', {
-  buildSW: {
-    swSrc: 'custom-sw.js',
-    swDest: 'custom-build/sw-cli-generated.js',
-    globDirectory: './custom-build',
-    globPatterns: ['**/*.{js,html}'],
+export default defineCliOptions('generate-sw', {
+  selfDestroying: {
+    selfDestroying: true,
+    swDest: 'sw-destroy.js',
   },
+
   generateSW: {
     globDirectory: './',
     globIgnores,
@@ -21,17 +18,28 @@ export default defineCliOptions('get-manifest', {
     swDest: 'sw.js',
     runtimeCaching,
   },
-  getManifest: {
-    globDirectory: './',
-    globIgnores,
+
+  buildSW: {
+    swSrc: 'src/sw.ts',
+    swDest: 'dist/sw.js',
+    globDirectory: './dist',
     globPatterns: ['**/*.{js,html}'],
+    globIgnores,
+    globStrict: false, // <== allows empty dist
   },
+
   injectManifest: {
     swSrc: 'custom-sw.js',
     swDest: 'custom-build/sw-cli-generated.js',
     injectionPoint: 'self.__WB_MANIFEST',
     globDirectory: './custom-build',
     globPatterns: ['**/*.{js,html}'],
+    globIgnores,
   },
 
+  getManifest: {
+    globDirectory: './',
+    globIgnores,
+    globPatterns: ['**/*.{js,html}'],
+  },
 })
